# Optimize index bounds check for immutable sorted set and list