--- a/src/libraries/System.Collections.Immutable/src/System/Collections/Immutable/ImmutableList_1.Node.cs
+++ b/src/libraries/System.Collections.Immutable/src/System/Collections/Immutable/ImmutableList_1.Node.cs
@@ -195,19 +195,9 @@ namespace System.Collections.Immutable
             internal ref readonly T ItemRef(int index)
             {
                 Requires.Range(index >= 0 && index < this.Count, nameof(index));
-
                 Debug.Assert(_left != null && _right != null);
-                if (index < _left._count)
-                {
-                    return ref _left.ItemRefUnchecked(index);
-                }
 
-                if (index > _left._count)
-                {
-                    return ref _right.ItemRefUnchecked(index - _left._count - 1);
-                }
-
-                return ref _key;
+                return ref ItemRefUnchecked(index);
             }
 
             internal ref readonly T ItemRefUnchecked(int index)