--- a/src/libraries/System.Collections.Immutable/src/System/Collections/Immutable/ImmutableSortedSet_1.Node.cs
+++ b/src/libraries/System.Collections.Immutable/src/System/Collections/Immutable/ImmutableSortedSet_1.Node.cs
@@ -265,12 +265,29 @@ namespace System.Collections.Immutable
 
                 if (index < _left._count)
                 {
-                    return ref _left.ItemRef(index);
+                    return ref _left.ItemRefUnchecked(index);
                 }
 
                 if (index > _left._count)
                 {
-                    return ref _right.ItemRef(index - _left._count - 1);
+                    return ref _right.ItemRefUnchecked(index - _left._count - 1);
+                }
+
+                return ref _key;
+            }
+
+            internal ref readonly T ItemRefUnchecked(int index)
+            {
+                Debug.Assert(_left != null && _right != null);
+
+                if (index < _left._count)
+                {
+                    return ref _left.ItemRefUnchecked(index);
+                }
+
+                if (index > _left._count)
+                {
+                    return ref _right.ItemRefUnchecked(index - _left._count - 1);
                 }
 
                 return ref _key;