@stephentoub 2021-05-26
Thanks.  Can you share benchmark results showing that the extra code required here to avoid that range check actually makes a meaningful difference? (Note that range check itself could be optimized slightly as well, to `Requires.Range((uint)index < (uint)this.Count, nameof(index));`, which would further reduce the impact of avoiding it.)