@L2 2021-05-26
Sure @stephentoub 

dotnet/performance diff with/without this change using --filter "\*ImmutableSortedSet\*" "\*ImmutableList\*"

summary:
better: 6, geomean: 1.202
total diff: 6

No Slower results for the provided threshold = 2% and noise filter = 25ns.

| Faster                                                                 | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ---------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Collections.IterateFor<String>.ImmutableSortedSet(Size: 512)    |      1.40 |          6618.82 |          4735.50 |         |
| System.Collections.IterateFor<Int32>.ImmutableSortedSet(Size: 512)     |      1.33 |          6613.09 |          4984.01 |         |
| System.Collections.IterateFor<Int32>.ImmutableList(Size: 512)          |      1.27 |          6485.47 |          5117.11 |         |
| System.Collections.IterateFor<String>.ImmutableList(Size: 512)         |      1.09 |          5579.58 |          5116.40 |         |
| System.Collections.IterateForEach<Int32>.ImmutableSortedSet(Size: 512) |      1.09 |         12384.04 |         11380.41 |         |
| System.Collections.ContainsTrue<Int32>.ImmutableSortedSet(Size: 512)   |      1.08 |         15290.25 |         14142.94 |         |