@L2 2021-05-26
@stephentoub thanks, I did encounter some places with these `Requires.Range` checks that could benefit from the optimization you listed. Should I place those changes in another PR or keep it in this one?