@L2 2021-05-28
Thanks @stephentoub , it looks like some recent commits have changed the perf numbers. Here's a breakdown of my investigation (Windows x64):

without my change + base commit = b443b8ccce174e (May 24, 2021) (The numbers I provided earlier were based on this)
with my change + base commit = b443b8ccce174e (May 24, 2021) (The numbers I provided earlier were based on this)
without my change + base commit = d43e88620fe57 (May 28, 2021)
with my change + base commit = d43e88620fe57 (May 28, 2021)

Today I cloned dotnet/runtime and created the four coreruns above and reran the microbenchmarks:

Here's an example of the commands I use:
```
py .\scripts\benchmarks_ci.py -c Release -f net6.0 --bdn-artifacts "C:\W052821\results\with_change" --corerun "C:\W052821\with_change\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.0\corerun.exe" --filter "*ImmutableSortedSet*" "*ImmutableList*"
```

```
py .\scripts\benchmarks_ci.py -c Release -f net6.0 --bdn-artifacts "C:\W052821\results\without_change" --corerun "C:\W052821\without_change\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.0\corerun.exe" --filter "*ImmutableSortedSet*" "*ImmutableList*"
```

then I diff the results using dotnet/performance's resultsComparer:
```
dotnet run --base "C:\W052821\results\without_change\" --diff "C:\W052821\results\with_change\"--threshold 2% --noise 25ns --full-id
```

--------------------------

Results:

without my change + base commit = b443b8ccce174e (May 24, 2021)
with my change + base commit = b443b8ccce174e (May 24, 2021) [DIFF]

summary:
better: 7, geomean: 1.191
worse: 1, geomean: 1.031
total diff: 8

| Slower                                                                  | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ----------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Collections.IterateForEach<String>.ImmutableSortedSet(Size: 512) |      1.03 |         22179.70 |         22860.75 |         |

| Faster                                                                        | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ----------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Collections.IterateFor<Int32>.ImmutableSortedSet(Size: 512)            |      1.41 |          6615.04 |          4681.89 |         |
| System.Collections.IterateFor<Int32>.ImmutableList(Size: 512)                 |      1.34 |          6480.18 |          4850.76 |         |
| System.Collections.IterateFor<String>.ImmutableSortedSet(Size: 512)           |      1.23 |          6629.84 |          5384.70 |         |
| System.Collections.IterateFor<String>.ImmutableList(Size: 512)                |      1.17 |          5595.22 |          4800.09 |         |
| System.Collections.IterateForEach<Int32>.ImmutableSortedSet(Size: 512)        |      1.09 |         12401.40 |         11399.09 |         |
| System.Collections.ContainsTrue<Int32>.ImmutableSortedSet(Size: 512)          |      1.09 |         15318.19 |         14107.30 |         |
| System.Collections.ContainsTrueComparer<String>.ImmutableSortedSet(Size: 512) |      1.06 |        212987.67 |        200160.62 |         |

--------------------------

without my change + base commit = d43e88620fe57 (May 28, 2021)
with my change + base commit = d43e88620fe57 (May 28, 2021) [DIFF]

summary:
better: 6, geomean: 1.101
worse: 1, geomean: 1.083
total diff: 7

| Slower                                                               | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Collections.ContainsTrue<Int32>.ImmutableSortedSet(Size: 512) |      1.08 |         14120.79 |         15288.74 | several?|

| Faster                                                                      | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| --------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Collections.IterateFor<String>.ImmutableList(Size: 512)              |      1.25 |          6497.43 |          5215.18 |         |
| System.Collections.IterateFor<String>.ImmutableSortedSet(Size: 512)         |      1.15 |          5587.32 |          4839.56 |         |
| System.Collections.IterateForEach<Int32>.ImmutableSortedSet(Size: 512)      |      1.08 |         12242.48 |         11376.06 |         |
| System.Collections.IterateFor<Int32>.ImmutableList(Size: 512)               |      1.07 |          5592.33 |          5229.02 | several?|
| System.Collections.CtorFromCollection<String>.ImmutableSortedSet(Size: 512) |      1.05 |        273723.27 |        261787.71 |         |
| System.Collections.IterateForEach<String>.ImmutableList(Size: 512)          |      1.03 |         22647.78 |         21959.21 |         |

------------------

Notes:
Rerunning on a fresh clone and build with base = b443b8ccce174e (May 24, 2021) with/without my change show the numbers being consistent on my machine with the numbers I provided earlier.

Building with the new base = d43e88620fe57 (May 28, 2021) seems to have changed some of the perf numbers around. Some faster and some slower than before:

without my change + base commit = b443b8ccce174e (May 24, 2021)
without my change + base commit = d43e88620fe57 (May 28, 2021) [DIFF]

summary:
better: 5, geomean: 1.138
worse: 3, geomean: 1.100
total diff: 8

| Slower                                                                      | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| --------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Collections.IterateFor<String>.ImmutableList(Size: 512)              |      1.16 |          5595.22 |          6497.43 |         |
| System.Collections.CtorFromCollection<String>.ImmutableSortedSet(Size: 512) |      1.08 |        252765.37 |        273723.27 |         |
| System.Collections.IterateForEach<String>.ImmutableList(Size: 512)          |      1.06 |         21371.06 |         22647.78 |         |

| Faster                                                                        | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ----------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Collections.IterateFor<Int32>.ImmutableSortedSet(Size: 512)            |      1.21 |          6615.04 |          5447.31 |         |
| System.Collections.IterateFor<String>.ImmutableSortedSet(Size: 512)           |      1.19 |          6629.84 |          5587.32 |         |
| System.Collections.IterateFor<Int32>.ImmutableList(Size: 512)                 |      1.16 |          6480.18 |          5592.33 | several?|
| System.Collections.ContainsTrue<Int32>.ImmutableSortedSet(Size: 512)          |      1.08 |         15318.19 |         14120.79 | several?|
| System.Collections.ContainsTrueComparer<String>.ImmutableSortedSet(Size: 512) |      1.06 |        212987.67 |        201713.26 |         |

------------------

So this change still shows some decent results with the latest base commit = d43e88620fe57 (May 28, 2021), but you're right not nearly as good as the numbers on base commit base commit = b443b8ccce174e (May 24, 2021). I'll follow your recommendation.