@L2 2021-06-09
Latest .NET6 (commit ac87f000c90814a76f78) before/after this change:

summary:
better: 5, geomean: 1.160
worse: 2, geomean: 1.053
total diff: 7

| Slower                                                                  | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ----------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Collections.ContainsTrue<Int32>.ImmutableSortedSet(Size: 512)    |      1.08 |         14111.28 |         15277.27 |         |
| System.Collections.IterateForEach<String>.ImmutableSortedSet(Size: 512) |      1.02 |         22045.85 |         22582.00 |         |

| Faster                                                                 | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ---------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Collections.IterateFor<String>.ImmutableList(Size: 512)         |      1.24 |          6476.43 |          5206.49 |         |
| System.Collections.IterateFor<Int32>.ImmutableSortedSet(Size: 512)     |      1.23 |          6476.55 |          5246.25 |         |
| System.Collections.IterateFor<String>.ImmutableSortedSet(Size: 512)    |      1.16 |          5593.19 |          4840.51 | several?|
| System.Collections.IterateFor<Int32>.ImmutableList(Size: 512)          |      1.16 |          5595.04 |          4842.67 |         |
| System.Collections.IterateForEach<Int32>.ImmutableSortedSet(Size: 512) |      1.02 |         11791.16 |         11525.10 |         |